==> analysis_Phermanns/06_fst_phase_only.R <==
# fst

library(parallel)
library(foreach)
library(doParallel)
library(dartR)
# setup dir info --------------------------------------
phaselength <- c(1, 2, 2, 4, 2, 4, 2, 2)
simdir <- '/data/scratch/emily/simulations/genlights/'
fstdir <- '/data/scratch/emily/simulations/fst/'
condir <- '/data/scratch/emily/simulations/meta/'
dir.create(fstdir)
sim_list <- list.files('/data/scratch/emily/simulations/genlights/')


# setup cores -----------------
ncores <- 9
cl <- parallel::makeCluster(ncores)
doParallel::registerDoParallel(cl)
tofst <- 20:27

# run fst ---------------------
system.time({
for(i in 1:length(sim_list)){
filename <- sim_list[i]
fstfilename <- sub('\\.rds', '_fst.rds', filename)
fst_list <- list.files('/data/scratch/emily/simulations/fst/')
if(!fstfilename %in% fst_list){
  print(paste('calculate fst for simulation', i, Sys.time()))
confilename <- paste0(condir, sub('.rds', '_conditions.csv', filename))
simAll <- readRDS(paste0(simdir, filename))

print('finish loading file...')
  simFst <- foreach(k = tofst) %dopar% {
    x <- TRUE
    while(x){
      fst <- dartR::gl.fst.pop(simAll[[k]], nboots = 1)
      if(is.matrix(fst)) x <- FALSE
    } # end while
    fst
  } # end foreach
  print(paste('finish calculating fsts...', Sys.time()))

saveRDS(simFst, paste0(fstdir, fstfilename))

conditionsx <- read.csv(confilename)
#conditionsx$fst <- sapply(simFst[1:19], mean, na.rm = T)
#conditionsx$fstphase <- rep(sapply(simFst[20:27], mean, na.rm = T), phaselength)
conditionsx$fst <- rep(sapply(simFst, mean, na.rm = T), phaselength)
conditionsx$fstphase <- rep(sapply(simFst, mean, na.rm = T), phaselength)
write.csv(conditionsx, confilename, row.names = FALSE)
}# end if
}# end for
})# end sys time
stopCluster(cl)
